requirements.txt
streamlit==1.39.0
pillow==10.2.0
python-docx==1.1.2
